+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Form can render a caption above non-remote forms.
+- Table can render a caption.
+- Table can be rendered with a top margin which can be in the range 0 to 7.
 ### Changed
 ### Fixed